hit_dirs is a pair in one word, and a byte lost half of it
`uint8_t d = global.hit_dirs[i]` where the field is a uint16_t and both
halves are read: `d & 0xff` is the horizontal and `d >> 8` the vertical,
and on a byte the second is always zero - so a ball hitting two bricks at
once stopped reversing vertically. Mine, from the batch that narrowed
every local called `d`, where I decided this one was "a direction not an
offset" and made it a byte on that reasoning without reading the two
lines under it.

1ac2:2599 settles the width: `mov dx, word ptr [di]` then `mov [si+15],
dh`. It is DX, split into DH and DL - the same shape as io_get_key's AX,
which stays a word for the same reason.

Two rules go in with it, both of which came from the same class of
mistake.

Rule 9, masks: `& 0xff` and `& 0xffff` are either the machine wrapping a
register or the type doing it twice, and the report splits them. A mask
on an assignment into a variable of that width is redundant; one in a
comparison or on a call argument whose parameter is wider is the wrap.
Only the outermost operation counts - `(ticks & 0xff) + (ticks >> 8)`
splits a word before adding the halves, and the first version of the rule
wrongly called that redundant.

Rule 10, casts to the width the target already has - 56 of them, from
`b->anchor_x = (uint8_t)(b->x - 1)` where both are bytes. It looks the
width up rather than trusting the name, so a cast that really narrows is
not reported. particle_height's keeps its cast: the assignment would do
the same truncation, but the comment above it turns on that truncation
being visible.

And rule 9 found what it was built for. Twelve routines took a coordinate
as uint16_t while verify.c's dispatch shows the original passing AL, BL
or CL - probe_cell_at, ball_draw, pixel_xor, shot_xor, sprite_shift_draw,
xor_sprite_20x16, xor_sprite_16xn, xor_sprite_16x7, draw_sprite_20x6,
brick_11_after, cell_hole_draw and draw_anim_cell. The ten `& 0xff` at
their call sites were the symptom. ending_plot keeps its words, because
1ac2:5add really is handed CX and DX.

Checked: ball_bricks, ball_after, ball_step, probe_cell_at, ball_draw,
ball_paddle, ball_redraw, brick_animated, score_add, draw_sprite_20x6,
laser_fire, entity_capsule, entity_popup, entity_hatch, entity_sparkle,
entity_crumble, bonus_spawn, bonus_update and the eight blitters all
agree with the original, and 2,500 frames on panel 8, 2,000 on level 11,
1,500 on level 50, 1,200 on panel 10 and 1,500 from the menu are
identical throughout.

+++ src/game.h
@@ -1649,7 +1649,7 @@ void plot_pixel_xor(uint16_t x, uint16_t y, uint8_t colour);
 void menu_particles_tick(void);   /* 1ac2:53c2 */
 void menu_banner_tick(void);      /* 1ac2:50df */
 void banner_shift(void);          /* 1ac2:5140 */
-void brick_11_after(uint16_t x, uint16_t y);  /* 1ac2:4c4b */
+void brick_11_after(uint8_t x, uint8_t y);  /* 1ac2:4c4b */
 uint16_t particle_random(uint16_t ax, uint16_t ticks, uint16_t limit); /* 1ac2:5448 */
 uint16_t particle_init(particle_t *p, uint16_t ax_in);  /* 1ac2:548a */
 void menu_arrow(void);            /* 1ac2:490d */
@@ -1674,7 +1674,7 @@ void cheat_match(uint8_t c);/* 1ac2:5171 */
 void io_cga_mode(uint32_t v);
 void io_cga_colour(uint32_t v);
 void employee_enter(void);        /* 1ac2:4ae0 */
-void cell_hole_draw(uint16_t x, uint16_t y);  /* 1ac2:4cc1 */
+void cell_hole_draw(uint8_t x, uint8_t y);  /* 1ac2:4cc1 */
 void screen_unstash(void);        /* 1ac2:4c13 */
 void border_setup(void);          /* 1ac2:4f73 */
 int32_t  tall_sprite(const uint8_t **si, uint16_t di); /* 1ac2:538d */
@@ -1724,7 +1724,7 @@ uint16_t draw_text(const char *src, uint16_t count, uint16_t di); /* 1ac2:10d1 *
 void level_draw(void);            /* 1ac2:1c4f */
 void walker_draw(uint16_t x);     /* 1ac2:1e50 */
 void walker_step(uint16_t x);     /* 1ac2:1e23 */
-void ball_draw(const void *rows, uint16_t x, uint16_t y);    /* 1ac2:2881 */
+void ball_draw(const void *rows, uint8_t x, uint8_t y);    /* 1ac2:2881 */
 int32_t  ball_redraw(ball_t *b);  /* 1ac2:2827 */
 int32_t  ball_on_paddle(ball_t *b); /* 1ac2:2e1e */
 void read_new_key(uint16_t which);  /* 1ac2:1614 */
@@ -1733,14 +1733,14 @@ void ball_after(ball_t *b);   /* 1ac2:247f */
 int32_t  ball_after_endgame(ball_t *b);  /* 1ac2:45a1 */
 void ball_bricks(ball_t *b);  /* 1ac2:254d */
 void brick_hit(hit_t *hit, uint8_t *cell, ball_t *ball);
-void xor_sprite_16xn(uint16_t x, uint16_t y, const uint8_t *src, uint16_t rows); /* 1ac2:40f2 */
+void xor_sprite_16xn(uint8_t x, uint8_t y, const uint8_t *src, uint16_t rows); /* 1ac2:40f2 */
 void brick_1(hit_t *hit, ball_t *ball);     /* 1ac2:28cb */
 void brick_2(hit_t *hit, ball_t *ball);     /* 1ac2:2985 */
 void brick_3(hit_t *hit, ball_t *ball);     /* 1ac2:2a3f */
 void brick_solid(hit_t *hit, ball_t *ball);       /* 1ac2:3221 */
 void brick_animated(hit_t *hit, ball_t *ball);   /* 1ac2:2ccd */
 void entity_anim_brick(ent_brick_t *a);                 /* 1ac2:3abf */
-void draw_anim_cell(const anim_sprite_t *sprite, uint16_t x, uint16_t y); /* 1ac2:3bac */
+void draw_anim_cell(const anim_sprite_t *sprite, uint8_t x, uint8_t y); /* 1ac2:3bac */
 void brick_5(hit_t *hit, ball_t *ball);     /* 1ac2:2a73 */
 void brick_6(hit_t *hit, ball_t *ball);     /* 1ac2:2ab4 */
 void brick_7(hit_t *hit, ball_t *ball);     /* 1ac2:2af5 */
@@ -1753,7 +1753,7 @@ void entity_ball_arrive(ent_anim_t *a); /* 1ac2:36a1 */
 void entity_cells_timer(ent_cells_t *a); /* 1ac2:36f6 */
 void brick_10(hit_t *hit, ball_t *ball);    /* 1ac2:2c59 */
 void brick_11(hit_t *hit, ball_t *ball);    /* 1ac2:2d68 */
-void xor_sprite_16x7(uint16_t x, uint16_t y, const uint8_t *src); /* 1ac2:3b64 */
+void xor_sprite_16x7(uint8_t x, uint8_t y, const uint8_t *src); /* 1ac2:3b64 */
 void score_add(void);             /* 1ac2:413d */
 void extra_life(void);            /* 1ac2:318b */
 void fill_column(uint16_t di, uint16_t value);  /* 1ac2:41b1 */
@@ -1775,7 +1775,7 @@ void draw_paddle_raw(const uint8_t *src);/* 1ac2:22a9 */
 void draw_paddle_shifted(const uint8_t *sprite); /* 1ac2:2187 */
 void ball_paddle(ball_t *b);  /* 1ac2:2316 */
 void laser_fire(void);            /* 1ac2:2ee3 */
-void probe_cell_at(uint16_t x, uint16_t y, hit_t *hit);    /* 1ac2:2755 */
+void probe_cell_at(uint8_t x, uint8_t y, hit_t *hit);    /* 1ac2:2755 */
 void play_teardown(void);         /* 1ac2:41d4 */
 void entity_call(entity_t *e);  /* the call at 1ac2:1b5e */
 void entity_capsule(ent_fall_t *f); /* 1ac2:3273 */
@@ -1787,8 +1787,8 @@ void entity_capsule_frames(ent_fall_t *f, uint16_t table_ptr);
 void entity_ball_hold(ent_anim_t *a); /* 1ac2:37e0 */
 void ball_place(ball_t *ball, uint16_t x, uint16_t y);
 void bonus_update(ent_sprite_t *s, uint16_t nx, uint16_t ny); /* 1ac2:3df1 */
-uint16_t pixel_xor(uint16_t x, uint16_t y);        /* 1ac2:30dd */
-void shot_xor(uint16_t x, uint16_t y);             /* 1ac2:306b */
+uint16_t pixel_xor(uint8_t x, uint8_t y);        /* 1ac2:30dd */
+void shot_xor(uint8_t x, uint8_t y);             /* 1ac2:306b */
 void bonus_hits_ball(const ent_sprite_t *s, const ball_t *ball);  /* 1ac2:3f20 */
 void entity_bonus(ent_anim_t *b);   /* 1ac2:39fa */
 void entity_unknown(entity_t *e);
@@ -1803,8 +1803,8 @@ void cell_set_three(ent_anim_t *a); /* 1ac2:3668 */
 void cells_restore(void);         /* 1ac2:36fb */
 
 void bonus_spawn(void);           /* 1ac2:3d95 */
-void xor_sprite_20x16(uint16_t x, uint16_t y, const uint8_t *src); /* 1ac2:406a */
-void sprite_shift_draw(uint16_t x, uint16_t y, const uint8_t *src); /* 1ac2:3f4f */
+void xor_sprite_20x16(uint8_t x, uint8_t y, const uint8_t *src); /* 1ac2:406a */
+void sprite_shift_draw(uint8_t x, uint8_t y, const uint8_t *src); /* 1ac2:3f4f */
 void entity_sparkle(ent_anim_t *a); /* 1ac2:3aee */
 void entity_crumble(ent_anim_t *a); /* 1ac2:3b2a */
 void entity_hatch(ent_hatch_t *h);   /* 1ac2:390d */
@@ -1839,7 +1839,7 @@ void panel_draw(void);            /* 1ac2:0b0b */
 void level_colours(void);         /* 1ac2:044b */
 void level_intro(void);           /* 1ac2:1eb9 */
 uint16_t draw_brick_row(uint16_t y);  /* 1ac2:2034 */
-void draw_sprite_20x6(uint16_t x, uint16_t y, const uint8_t *src); /* 1ac2:20b9 */
+void draw_sprite_20x6(uint8_t x, uint8_t y, const uint8_t *src); /* 1ac2:20b9 */
 void cell_special(uint8_t row, uint8_t col, uint16_t di); /* 1ac2:41e5 */
 void field_backdrop(uint16_t y);  /* 1ac2:1fc1 */
 void life_lost(void);             /* 1ac2:0735 */
